PCB SENTINEL PRO/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
ultralytics==8.1.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python==4.9.0.80
numpy>=1.24.0
pillow==10.2.0
pandas==2.2.0
requests==2.31.0
python-multipart==0.0.6
pyyaml>=6.0
